pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "milodex"
version = "0.1.0"
description = "A personal, research-led trading system that discovers, validates, deploys, and monitors strategies."
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
dependencies = [
    "alpaca-py>=0.35.0",
    "pandas>=2.0",
    "pyarrow>=15.0",
    "python-dotenv>=1.0",
    "pytz>=2024.1",
    "PyYAML>=6.0",
]

[project.scripts]
milodex = "milodex.cli.main:main"

[project.optional-dependencies]
dev = [
    "ruff",
    "pytest",
]

[tool.ruff]
src = ["src"]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
markers = [
    "integration: marks tests that hit real APIs (deselect with '-m \"not integration\"')",
]
